requirements.txt
# EMG dependencies
# install and create a virtual environment
# run pip install -r requirements

# deployment

gunicorn==20.1.0
mysqlclient==2.1.1
mysql-connector-python~=8.0.23
sqlparse==0.4.3
whitenoise==6.4.0
requests==2.28.2

yamjam==0.1.7
# python 3.4
PyYAML==6.0

# log handler
concurrent-log-handler~=0.9.22

Django==3.2.18
djangorestframework==3.12
django-filter==23.1
djangorestframework-jwt~=1.11.0
django-cors-headers==3.14.0
djangorestframework-jsonapi==4.2.1
cx_Oracle~=6.2.1

djangorestframework-csv==2.1.1

# schema
drf-spectacular==0.26.0

# mongo
mongoengine==0.27.0
pymongo==4.3.3
django-rest-framework-mongoengine==3.4.1

# assembly viewer
pysam==0.21.0

# sourmash search
celery[redis]==5.2.7

# my-sql utils
django-mysql==4.3.0
